Added Content to Footer
Also positioned content horizontally, and aligned it with eachother
vertically (github links showing on same line as copyright statement
rather than underneath). Height of footer was adjusted and set in
app.component.scss, but not sure if it's a permament change. Space not
currently needed. Minor changes to color styles of header and footer as
well to begin following more of a color theme.